Finish review fixes for renumber and wrap tests
Replace the silent `write!` header assembly in the footnote renumberer
with infallible string construction, and stop parsing numeric candidates
through panic-prone capture indexing.

Move the `wrap_text` regression coverage out of the oversized
`src/wrap/tests.rs` file into the existing dedicated integration test
surface, and strengthen the Unicode-indent case so it proves display
width rather than byte length.
